+++ src/common/jwt-configuration-parameters.ts
@@ -1,0 +1,4 @@
+export interface JwtConfigurationParameters {
+  saltRounds: number,
+  secretKey: string
+}