+++ tests/mocks/koal-app-mock.ts
@@ -1,0 +1,7 @@
+import { Configuration, ConfigurationParameters, KoalApp } from "../../src";
+
+export function getKoalAppMock(configurationParameters: ConfigurationParameters) {
+  return KoalApp.getInstance(new Configuration({
+    ...configurationParameters
+  }));
+};